requirements.txt
streamlit==1.8.1
torch==1.11.0
torchvision==0.12.0
Pillow
matplotlib
